clean up

+++ includes/domains.php
@@ -11,10 +11,9 @@
         require_once( INCLUDE_DIR."/imap.inc.php" );
 
         $sel = chks2sql( 'domain_id' );
-        if ( $sel != "" ) {
+        if ( $sel ) {
             sql_query( "SELECT * FROM cyrup_accounts WHERE ".$sel );
-            while ( $row = sql_fetch_array() )
-            	cimap_deletemailbox( $row['account'] );
+            while ( $row = sql_fetch_array() ) cimap_deletemailbox( $row['account'] );
             sql_query( "DELETE FROM cyrup_accounts WHERE ".$sel );
             sql_query( "DELETE FROM cyrup_aliases WHERE ".$sel );
             sql_query( "DELETE FROM cyrup_default_rcpt WHERE ".$sel );
@@ -22,8 +21,7 @@
         }
 
         $sel = chks2sql( 'id' );
-        if ( !empty($sel) )
-            sql_query( 'DELETE FROM cyrup_domains WHERE '.$sel );
+        if ( !empty($sel) ) sql_query( 'DELETE FROM cyrup_domains WHERE '.$sel );
         if ( defined('DOMAIN_EXPORT_FILE') AND DOMAIN_EXPORT_FILE != '' )
             sql_export( 'SELECT domain FROM cyrup_domains WHERE enabled=1' , DOMAIN_EXPORT_FILE );
         if ( defined('SYSTEM_ALIASES') AND SYSTEM_ALIASES != '' ) 
@@ -41,7 +39,7 @@
     if ( $_SESSION['USER'] == ADMIN_USER ) {
 	print "<form name=form method=POST action='".BASE_URL."/?admin&m=domains'>\n";
 	print "<input type=hidden name=action value=''>\n";
-    };
+    }
     print "<table width=100% border=0 cellpadding=0 cellspacing=0>\n";
     dotline( $colspan );
 
@@ -52,8 +50,7 @@
     html_th( "accounts_max", "Accounts / Max", 'Accounts in use / Maximum allowed in domain' );
     html_th( "aliases_max", "Aliases / Max", 'Aliases in use / Maximum allowed in domain' );
     html_th( "quota", "Quota current / Max ", 'Quota in use / Maximum allowed in domain' );
-    if ( MAILBOX_STYLE == "USERSUFFIX" )
-        html_th( "account_suffix", "Acc. suffix" );
+    if ( MAILBOX_STYLE == "USERSUFFIX" ) html_th( "account_suffix", "Acc. suffix" );
     html_th( "aliased_to", "Def. email", "Unknown recipient's catcher" );
     html_th( "owner", "Owner" );
     html_th( "info", "Info" );
@@ -61,17 +58,11 @@
 
     dotline( $colspan );
     print "<tr class=highlight><td colspan=".$colspan." align=center>";
-    if ( $_SESSION['USER'] == ADMIN_USER ) 
-	print "<a href='?admin&m=domainform' class=button>[ Add new ]</a>";
-    else
-	 print "&nbsp;";
+    print $_SESSION['USER'] == ADMIN_USER ? "<a href='?admin&m=domainform' class=button>[ Add new ]</a>" : "&nbsp;";
     print "</td></tr>\n";
     dotline( $colspan );
 
-    $query = "SELECT id FROM cyrup_domains a 
-			    LEFT JOIN cyrup_default_rcpt b ON a.id=b.domain_id ".
-			    rights2sql(1,"a.id"). 
-			   " ORDER BY ".$order_by;
+    $query = "SELECT id FROM cyrup_domains a LEFT JOIN cyrup_default_rcpt b ON a.id=b.domain_id ".rights2sql(1,"a.id")." ORDER BY ".$order_by;
     $domains_res = sql_query( $query );
     $i = 0;
     while ( $row = sql_fetch_array( $domains_res ) ) {
@@ -87,18 +78,16 @@
 	print "</td>\n<td align='center'>".($domain_row['enabled'] ? 'Y' : 'N')."</td>\n";
         print "<td align=center>&nbsp;".$domain_row['accounts_cur']."/".$domain_row['accounts_max']."</td>\n";
 	print "<td align=center>&nbsp;".$domain_row['aliases_cur']."/".$domain_row['aliases_max']."</td>\n";
-        print "<td align=center>&nbsp;".kb2mb( $domain_row['quota_cur'] )."/"
-            .( $domain_row['quota'] ? kb2mb( $domain_row['quota'] ) : "no-quota" )."</td>\n";
-        if ( MAILBOX_STYLE == "USERSUFFIX" )
-            print "<td align=center>&nbsp;".$domain_row['account_suffix']."</td>\n";
+        print "<td align=center>&nbsp;".kb2mb( $domain_row['quota_cur'] )."/".($domain_row['quota'] ? kb2mb( $domain_row['quota'] ) : "no-quota" )."</td>\n";
+        if ( MAILBOX_STYLE == "USERSUFFIX" ) print "<td align=center>&nbsp;".$domain_row['account_suffix']."</td>\n";
         print "<td align=center>&nbsp;".$domain_row['default_rcpt']."</td>\n";
 	if ( $owner = get_domain_owner($row['id']) ) {
-	    sql_query( "SELECT username FROM cyrup_admins WHERE id = '".$owner."'" );
+	    sql_query( "SELECT username FROM cyrup_admins WHERE id = ".sql_escape($owner) );
 	    if ( $_SESSION['USER'] == ADMIN_USER )
 		$owner = "<a href='?admin&m=adminform&id=".$owner."'>".sql_fetch_variable()."</a>";
 	    else
 		$owner = "<a href='?admin&m=service'>".sql_fetch_variable()."</a>";
-	};
+	}
 	print "<td align=center>&nbsp;".$owner."</td>\n";
         print "<td align=right>&nbsp;".$domain_row['info']."</td>\n";
         print "</tr>\n";
@@ -106,10 +95,7 @@
     };
     print "</table>\n";
     if ( $_SESSION['USER'] == ADMIN_USER ) {
-	if ( $i >= 1 ) 
-	    print "<br><br>\n".delete_selected_box();
+	if ( $i >= 1 ) print "<br><br>\n".delete_selected_box();
 	print "</form>\n";
-    };
+    }
     print_footer();
-
-?>
